backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.2
boto3==1.12.48
botocore==1.15.48
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-location-field==2.1.0
django-mysql==3.3.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
dotty-dict==1.2.1
drf-yasg==1.17.1
gitdb==4.0.4
GitPython==3.1.1
idna==2.8
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pandas==0.25.3
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pyotp==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
quotes==0.0.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
setuptools-scm==3.5.0
simplejson==3.17.0
six==1.13.0
smmap==3.0.2
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
s2qlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
